Respond to PR feedback.
