+++ src/System.Net.Http/tests/FunctionalTests/HttpClientHandlerTest.cs
@@ -888,7 +888,7 @@ namespace System.Net.Http.Functional.Tests
         }
 
         [Fact]
-        [OuterLoop]
+        [OuterLoop] // Test uses azure endpoint.
         public async Task GetAsync_CredentialIsNetworkCredentialUriRedirect_StatusCodeUnauthorized()
         {
             HttpClientHandler handler = CreateHttpClientHandler();
@@ -908,7 +908,7 @@ namespace System.Net.Http.Functional.Tests
         }
 
         [Fact]
-        [OuterLoop]
+        [OuterLoop] // Test uses azure endpoint.
         public async Task HttpClientHandler_CredentialIsNotCredentialCacheAfterRedirect_StatusCodeOK()
         {
             HttpClientHandler handler = CreateHttpClientHandler();
